+++ KickLib.Api.Unofficial/Extensions/IKickLibBuilder.cs
@@ -16,11 +16,11 @@ namespace KickLib.Api.Unofficial.Extensions
         /// <summary>
         ///     Adds client that uses Puppeteer to spin headless browser.
         /// </summary>
-        IServiceCollection WithPuppeteerClient();
+        IServiceCollection WithPuppeteerClient(BrowserSettings browserSettings = null);
 
         /// <summary>
         ///     Adds optimized client that uses shared Puppeteer browser instance with multi-user support.
         /// </summary>
-        IServiceCollection WithOptimizedPuppeteerClient(BrowserSettings browserSettings = null);
+        IServiceCollection WithSessionPuppeteerClient(BrowserSettings browserSettings = null);
     }
 }